@@ -2,7 +2,7 @@
     "name": "vscode-appcenter",
     "displayName": "App Center",
     "description": "Visual Studio App Center tools and services.",
-    "version": "0.1.11",
+    "version": "0.2.0",
     "publisher": "vsmobile",
     "icon": "images/icon.png",
     "engines": {
@@ -286,7 +286,7 @@
         "vscode:prepublish": "gulp",
         "compile": "tsc -p ./",
         "watch": "tsc -watch -p ./",
-        "lint": "tslint --project tsconfig.json -e src/*.d.ts -t verbose",
+        "lint": "eslint --ext .ts ./src --fix",
         "postinstall": "node ./node_modules/vscode/bin/install",
         "test": "./node_modules/.bin/mocha --reporter spec",
         "integrationTest": "node ./node_modules/vscode/bin/test"
@@ -295,14 +295,17 @@
         "@types/mocha": "^2.2.42",
         "@types/node": "^7.0.62",
         "@types/rimraf": "^2.0.2",
-        "gulp": "^3.9.1",
+        "@typescript-eslint/eslint-plugin": "^5.37.0",
+        "@typescript-eslint/parser": "^5.37.0",
+        "eslint": "8.22.0",
+        "eslint-config-prettier": "^8.5.0",
+        "gulp": "^4.0.2",
         "gulp-sourcemaps": "^2.6.4",
         "gulp-tslint": "^8.1.3",
-        "gulp-typescript": "^4.0.2",
+        "gulp-typescript": "^2.12.2",
         "mock-fs": "^4.5.0",
-        "tslint": "^5.9.1",
-        "tslint-microsoft-contrib": "^5.0.3",
-        "typescript": "^2.8.3",
+        "prettier": "^2.7.1",
+        "typescript": "^4.8.3",
         "vscode": "^1.1.21"
     },
     "dependencies": {
@@ -311,9 +314,9 @@
         "date-fns": "^1.29.0",
         "del": "^3.0.0",
         "event-stream": "3.3.4",
-        "git-url-parse": "^8.3.1",
+        "git-url-parse": "^13.1.0",
         "gradle-to-js": "^1.1.0",
-        "gulp-mocha": "^5.0.0",
+        "gulp-mocha": "^8.0.0",
         "jsonwebtoken": "^8.2.1",
         "lodash": "^4.17.10",
         "mkdirp": "^0.5.1",
@@ -322,22 +325,31 @@
         "node-fetch": "^2.1.2",
         "node-jsxml": "^0.9.0",
         "node-noop": "^1.0.0",
-        "open": "0.0.5",
+        "open": "^8.4.0",
         "opener": "^1.4.3",
         "plist": "^3.0.1",
         "properties": "^1.2.1",
         "rimraf": "^2.6.2",
         "run-sequence": "^2.2.1",
         "rx-lite": "^4.0.8",
-        "sandbox": "^0.8.6",
         "semver": "^5.5.0",
         "should": "^13.2.1",
-        "simple-git": "^1.92.0",
+        "simple-git": "^3.14.1",
         "sinon": "^4.5.0",
         "temp": "^0.8.3",
         "test": "^0.6.0",
         "ts-md5": "^1.2.4",
         "xml2js": "^0.4.19",
         "yazl": "^2.4.3"
+    },
+    "prettier": {
+        "printWidth": 120,
+        "tabWidth": 4,
+        "singleQuote": true,
+        "trailingComma": "all",
+        "bracketSpacing": true,
+        "parser": "typescript",
+        "semi": true,
+        "useTabs": false
     }
 }
